opt: make we2 uart dma work smoother

+++ core/utils/el_ringbuffer.hpp
@@ -16,11 +16,11 @@ class lwRingBuffer {
         buf[tail] = c;
         tail      = (tail + 1) % len;
     }
-    size_t put(const char* str, int slen) {
+    size_t put(const char* str, size_t slen) {
         if (slen > free()) {
             slen = free();
         }
-        for (int i = 0; i < slen; i++) {
+        for (size_t i = 0; i < slen; i++) {
             put(str[i]);
         }
         return slen;
@@ -35,11 +35,11 @@ class lwRingBuffer {
         head   = (head + 1) % len;
         return c;
     }
-    size_t get(char* str, int slen) {
+    size_t get(char* str, size_t slen) {
         if (slen > size()) {
             slen = size();
         }
-        for (int i = 0; i < slen; i++) {
+        for (size_t i = 0; i < slen; i++) {
             str[i] = get();
         }
         return slen;
